refatorada funcao que cria o texto da tooltip, refatorados estilos da tooltip
